drafted registration.py for point cloud registration
