+++ README.md
@@ -1,3 +1,3 @@
-This repulsitory is my individual project in Imperial College London. Some of my notes can be foundin the `markdowns` directory.  
+This repulsitory is my individual project in Imperial College London. Some of my notes can be found in the `markdowns` directory, I recommend checking out the `memo.md` file to get more idea on what I have done in the project. Drafts of my work can be found in the `notebooks` directory, and I am currently working my projects in the python files.
